[zsh] Remove zplug
Because almost all of my configuration was local, I was using this as a
complicated way to source files and download binaries from GitHub.
Unfortunately, the GitHub part stopped working due to GitHub changes and
zplug seems to be abandonded for the time being.

Simplify things by sourcing files directly from .zshrc. I expect zplug
incidentally did some things that I found useful, so I'll need to use
this for a while to see what else is missing.

I haven't replaced the binary installation part yet - I haven't decided
if this tool should handle that or if they should be installed manually.
I only regularly use ripgrep and jq of the installed tools.